requirements.txt
Flask==2.2.2
pandas==1.5.3
python-pptx==0.6.21
PyPDF2==3.0.1
matplotlib==3.6.2
pytest==7.2.0
Werkzeug==2.2.2
numpy==1.26.4